user/src/main/java/com/qaportal/user/dto/UserDTO.java
package com.qaportal.user.dto;

import java.util.Date;

public class UserDTO {
	
	private int userid;

	private String firstname;

	private String lastname;

	private String emailid;

	private String password;

	private Date createdate;

	public UserDTO(int userid, String firstname, String lastname, String emailid, String password, Date createdate) {
		super();
		this.userid = userid;
		this.firstname = firstname;
		this.lastname = lastname;
		this.emailid = emailid;
		this.password = password;
		this.createdate = createdate;
	}

	public UserDTO() {
	}

	public int getUserid() {
		return userid;
	}

	public void setUserid(int userid) {
		this.userid = userid;
	}

	public String getFirstname() {
		return firstname;
	}

	public void setFirstname(String firstname) {
		this.firstname = firstname;
	}

	public String getLastname() {
		return lastname;
	}

	public void setLastname(String lastname) {
		this.lastname = lastname;
	}

	public String getEmailid() {
		return emailid;
	}

	public void setEmailid(String emailid) {
		this.emailid = emailid;
	}

	public String getPassword() {
		return password;
	}

	public void setPassword(String password) {
		this.password = password;
	}

	public Date getCreatedate() {
		return createdate;
	}

	public void setCreatedate(Date createdate) {
		this.createdate = createdate;
	}
	
	
}
